LayoutTests/media/modern-media-controls/fullscreen-button/fullscreen-button-expected.txt

Testing the FullscreenButton class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS enterFullscreenButton.element.localName is "button"
PASS enterFullscreenButton.element.classList.contains("icon") is true
PASS enterFullscreenButton.element.classList.contains("fullscreen") is true
PASS enterFullscreenButton.iconName is Icons.EnterFullscreen
PASS exitFullscreenButton.iconName is Icons.ExitFullscreen
PASS enterFullscreenButton.element.style.webkitMaskImage.includes("macOS/enter-fullscreen@") is true
PASS iOSEnterFullscreenButton.element.style.webkitMaskImage.includes("iOS/enter-fullscreen@") is true
PASS exitFullscreenButton.element.style.webkitMaskImage.includes("macOS/exit-fullscreen@") is true
PASS successfullyParsed is true

TEST COMPLETE
